Refactoring Template container into wrapper and container
