+++ src/web/components/Personalities/Type.js
@@ -10,7 +10,7 @@ import {
 } from 'reactstrap';
 import { Link } from 'react-router-dom';
 
-import TemplateContainer from '../Templates/TemplateContainer';
+import { TemplateContainer } from '../Templates/Templates';
 import ErrorMessages from '../../../constants/errors';
 import Loading from '../Loading';
 import Error from '../Error';
