+++ src/web/styles/_globals.scss
@@ -2,3 +2,7 @@ $purple: #473BE7;
 $white: #fff;
 
 $typeLight: $white;
+
+$s-space: 8px;
+$m-space: 16px;
+$l-space: 32px;
